+++ includes/Core/REST_API/Data_Request.php
@@ -8,7 +8,7 @@
  * @link      https://sitekit.withgoogle.com
  */
 
-namespace Google\Site_Kit\Core\Util;
+namespace Google\Site_Kit\Core\REST_API;
 
 /**
  * Class Data_Request
