+++ tests/phpunit/integration/Core/Modules/ModuleTest.php
@@ -11,7 +11,7 @@
 namespace Google\Site_Kit\Tests\Core\Modules;
 
 use Google\Site_Kit\Context;
-use Google\Site_Kit\Core\Util\Data_Request;
+use Google\Site_Kit\Core\REST_API\Data_Request;
 use Google\Site_Kit\Tests\TestCase;
 
 /**
